requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-Bcrypt==1.0.1
Flask-Sock==0.7.0
Flask-Caching==2.0.1
Flask-Swagger-Ui==5.21.0
Flask-Cors==6.0.1
flasgger==0.9.7.1
redis==6.4.0
Celery==5.5.3
requests==2.32.5
reportlab==4.4.4
python-dotenv==1.1.1
pytest==8.4.2
pytest-flask==1.3.0
pytest-asyncio==1.2.0
PyPDF2==3.0.1
cryptography==46.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
PyMySQL==1.1.2
pika==1.3.2
eventlet==0.40.3
aio-pika==9.5.7